// src/modulos/ventas/model/factura-electronica/interfaces/de-establecimiento.interface.ts
export interface DEEstablecimientoInterface {
    codigo: string;
    direccion: string;
    numeroCasa: string;
    departamento: number;
    departamentoDescripcion: string;
    distrito: number,
    distritoDescripcion: string,
    ciudad: number;
    ciudadDescripcion: string;
    telefono: string;
    email: string;
    denominacion: string;
}